moved navigator and locator into disk.py
